cleaned up code and removed debug prints

+++ sudokil/core/src/com/custardgames/sudokil/entities/ecs/processes/LiftProcess.java
@@ -109,11 +109,11 @@ public class LiftProcess extends EntityProcess
 
 				if (deltaX != 0)
 				{
-					deltaX = (float) (-maxVelocity * (deltaX / Math.abs(deltaX)));
+					deltaX = -maxVelocity * (deltaX / Math.abs(deltaX));
 				}
 				if (deltaY != 0)
 				{
-					deltaY = (float)(-maxVelocity * (deltaY / Math.abs(deltaY)));
+					deltaY = -maxVelocity * (deltaY / Math.abs(deltaY));
 				}
 				liftedPosition.setX(positionX + deltaX);
 				liftedPosition.setY(positionY + deltaY);
